ASYNC-234 timeout2
timeout2 provides the same interface as timeout, but internally is
implemented rather differently. It returns a custom ReadPort which
doesn't add any references to the global timeouts-queue unless a
handler is register (via take!). The handler goes on the
timeouts-queue directly, and the NackableHandler protocol is used to
remove any handlers from the queue that become inactive before being
timedout.